requirements.txt
discord==1.7.3
youtube-dl==2021.6.6
requests==2.26.0
PyNaCl==1.4.0
discord.py==1.7.3
async-timeout==3.0.1
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.2
google-cloud-translate==3.3.2
google-trans-new==1.1.9
googleapis-common-protos==1.53.0
googletrans==3.1.0a0
grpcio==1.39.0
ffmpy==0.3.0

